requirements.txt
feedparser==6.0.12
beautifulsoup4==4.14.3
weasyprint==66.0
playwright==1.58.0
python-dotenv==1.2.1
jinja2==3.1.6
requests==2.32.3
icalendar==6.3.2

